straggler_mitigate/requirements.txt
numpy~=1.22.3
termcolor~=1.1.0
pytorch~=1.13.1
gym~=0.21.0
tensorboard~=2.6.0
pandas~=1.4.2
tqdm~=4.64.0
numba~=0.56.4
matplotlib~=3.6.2
scipy~=1.7.3
numba~=0.56.4
setuptools~=59.8.0
cython~=0.29.28